benefit and images controllers/repos

+++ src/main/java/com/exercise/mybnb/controller/ImageController.java
@@ -5,19 +5,14 @@ import com.exercise.mybnb.repository.ImageRepo;
 import com.exercise.mybnb.repository.PlaceRepo;
 import com.exercise.mybnb.utils.Utils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
 
 @RestController
 public class ImageController {
@@ -26,29 +21,28 @@ public class ImageController {
     @Autowired
     PlaceRepo placeRepo;
 
-    @GetMapping("/images")
+    @GetMapping("/images/names")
     public List<Image> getAllImages() {
         return imageRepo.findAll();
     }
 
-    @GetMapping("/images/{id}")
-    public Optional<Image> getImage(@PathVariable("id") int id) {
-        return imageRepo.findById(id);
-    }
-
     @GetMapping("/places/{pid}/images")
-    public Page<Image> getAllImagesByPlaceId(@PathVariable("pid") int pid, Pageable pageable) {
-        return imageRepo.findByPlaceId(pid, pageable);
+    public Optional<Set<Image>> getAllImages(@PathVariable("pid") int pid) {
+        return imageRepo.findByPlaceId(pid);
     }
 
     @PostMapping("/places/{pid}/images")
     public ResponseEntity<?> createImage(@PathVariable("pid") int pid,
-                                         MultipartFile[] files) {
+                                         @RequestParam("images") MultipartFile[] files) {
+        System.out.println("Posting images" + pid);
         if (files == null)
             throw new ResourceNotFoundException("Image file not found");
         return placeRepo.findById(pid).map(place -> {
+            System.out.println("Place? " + place.getAddress());
+            System.out.println("Mults? " + files.length);
             Utils.makeDir(pid);
             for (MultipartFile file : files) {
+                System.out.println("image got: " + file.getOriginalFilename());
                 Image image = new Image();
                 image.setFilename(file.getOriginalFilename());
                 image.setPlace(place);
@@ -64,32 +58,32 @@ public class ImageController {
         }).orElseThrow(() -> new ResourceNotFoundException("PlaceID " + pid + " not found"));
     }
 
-    //PUT????
-
-    @DeleteMapping("/places/{pid}/images/{id}")
-    public ResponseEntity<?> deleteImage(@PathVariable("pid") int pid,
-                                         @PathVariable("id") int id) {
-        return imageRepo.findByIdAndPlaceId(id, pid).map(image -> {
-            //ALSO DELETE THE IMAGE
-            Utils.deletePlaceImage(pid+"/" + image.getFilename());
-            imageRepo.delete(image);
-            return ResponseEntity.ok().build();
-        }).orElseThrow(() -> new ResourceNotFoundException("Image not found with id "
-                + id + " and PlaceID " + pid));
+    @GetMapping("/images/{id}")
+    public ResponseEntity<byte[]> getImage(@PathVariable("id") int id) {
+        return imageRepo.findById(id).map(image -> {
+            byte[] imBytes = new byte[0];
+            try {
+                imBytes = Utils.getImageBytes(image.getPlace(), image.getFilename());
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+            return ResponseEntity.ok()
+                    .contentType(MediaType.IMAGE_JPEG)
+                    .body(imBytes);
+        }).orElseThrow(() -> new ResourceNotFoundException("Image " + id + " not found"));
     }
 
-    @PostMapping("/random")
-    public Map<String, String> randomRequest(@RequestParam MultiValueMap<String, String> data){
-        System.out.println(data.get("key1").toString());
-        System.out.println(data.get("key2"));
-        System.out.println(data.get("key3"));
-
-        Map<String, String> resp = new HashMap<String, String>();
-        resp.put("KEY1", String.valueOf(data.get("key1")).toUpperCase());
-        resp.put("KEY2", String.valueOf(data.get("key2")).toUpperCase());
-        resp.put("KEY3", String.valueOf(data.get("key3")).toUpperCase());
-
-        return resp;
+    @DeleteMapping("/images/{id}")
+    public ResponseEntity<?> removeImage(@PathVariable("id") int id){
+        return imageRepo.findById(id).map(image -> {
+            String filepath = Utils.getMainPath() + "places/" + image.getPlace().getId() + "/" + image.getFilename();
+            if (Utils.deleteImage(filepath)){
+                imageRepo.delete(image);
+                return ResponseEntity.ok().build();
+            }
+            else
+                return ResponseEntity.notFound().build();
+        }).orElseThrow(() -> new ResourceNotFoundException("Image " + id + " not found"));
     }
 
 }
